Upgrade ion-java
Java SDK v2 dropped support for Ion, and it’s unsupported for v1.
The version of ion-java (software.amazon.ion:ion-java:1.0.2) bundled with aws-sdk-java is substantially behind the current version (com.amazon.ion:ion-java:1.9.6) and uses a different java package.
This causes customers using aws-sdk-java to get an out-dated version in their dependencies. It would be better to have an up-to-date version or no transitive dependency so that customers don't miss out on the bug fixes and  performance improvements ion-java has made since then.

This commit upgrades ion-java, adopting the most recent released version
with all of its bug fixes and performance improvements, as well as the
java package change.